refactor: Rename package ui to controller
